refactor: rename createProduct helper to createProductOrVariant
